不能接受前端请求，

+++ src/main/java/com/ebook/backend/dao/BookDao.java
@@ -3,7 +3,6 @@ package com.ebook.backend.dao;
 import com.ebook.backend.entity.Book;
 import com.ebook.backend.utils.messagegutils.Message;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface BookDao {
@@ -12,7 +11,7 @@ public interface BookDao {
     List<Book> getBooks();
 
     Message addBook(String isbn, String name, String type, String author, String description,
-                    String image, Integer num, Boolean state, BigDecimal price);
+                    String image, Integer num, Boolean state, Integer price);
     Book getBookById(Integer bookId);
 
     Message changeSale(Integer bookId, Integer purchaseNum);
