+++ src/main/java/com/ebook/backend/service/BookService.java
@@ -3,7 +3,6 @@ package com.ebook.backend.service;
 import com.ebook.backend.entity.Book;
 import com.ebook.backend.utils.messagegutils.Message;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -14,6 +13,6 @@ public interface BookService {
     List<Book> getBooks();
 
     Message addBook(String isbn , String name, String type , String author, String description, Integer num,
-                    String image, Boolean state, BigDecimal price);
+                    String image, Boolean state, Integer price);
 
 }
