+++ src/main/java/com/ebook/backend/serviceimpl/BookServiceImpl.java
@@ -7,7 +7,6 @@ import com.ebook.backend.utils.messagegutils.Message;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -29,7 +28,7 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public Message addBook(String isbn, String name, String type, String author, String description, Integer num, String image, Boolean state, BigDecimal price) {
+    public Message addBook(String isbn, String name, String type, String author, String description, Integer num, String image, Boolean state, Integer price) {
         return bookDao.addBook(isbn, name, type, author, description, image, num, state, price);
 
     }
